videos-upload/requirements.txt
django<3.3
django-filter
django-netfields
djangorestframework
drf_yasg
gunicorn
httptools

